chore: separate tests from chip file
