+++ zk_prover/src/chips/less_than/mod.rs
@@ -1,0 +1,2 @@
+pub mod less_than_vertical;
+mod tests;
