+++ zk_prover/src/chips/mod.rs
@@ -1,6 +1,6 @@
 pub mod aggregation;
-pub mod less_than_vertical;
 pub mod ecdsa;
+pub mod less_than;
 pub mod merkle_sum_tree;
 pub mod overflow;
 pub mod poseidon;
